+++ .env.example
@@ -8,20 +8,20 @@ Server_Port=8080
 Redis_Address=localhost:6379
 Environment=development
 AWS_ACCESS_KEY_ID=admin
-AwsSecretAccessKey=password
+AWS_SECRET_ACCESS_KEY=password
 S3Bucket=user-images
 S3Directory=story-images
 
-# docker cofig
-# MysqlDB_Connection_String=root:some_pass@tcp(host.docker.internal:3306)/userdb?charset=utf8mb4&parseTime=True&loc=Local
-# DynamoDB_Connection_String=http://localhost:4566
-# DynamoDB_Table=user-images
-# AWS_REGION=eu-central-1
-# Server_Host=0.0.0.0
-# Server_Port=8080
-# Redis_Address=host.docker.internal:6379
-# Environment=development
-# AWS_ACCESS_KEY_ID=admin
-# AwsSecretAccessKey=password
-# S3Bucket=user-images
-# S3Directory=story-images
+## docker cofig
+#MysqlDB_Connection_String=root:some_pass@tcp(host.docker.internal:3306)/userdb?charset=utf8mb4&parseTime=True&loc=Local
+#DynamoDB_Connection_String=http://docker.for.mac.host.internal:4566
+#DynamoDB_Table=user-images
+#S3Bucket=user-images
+#S3Directory=story-images
+#Server_Host=0.0.0.0
+#Server_Port=8080
+#Redis_Address=host.docker.internal:6379
+#Environment=development
+#AWS_REGION=eu-central-1
+#AWS_ACCESS_KEY_ID=admin
+#AWS_SECRET_ACCESS_KEY=password
