+++ Makefile
@@ -4,16 +4,6 @@ build: sanitize
 run-service: build
 	docker run -d --name=user-service --env-file .env -p 8080:8080 service-user
 
-build-local:
-	GOOS=linux GOARCH=arm64/v8 cd ./cmd/service-user && go build -a -o ../../dist/service
-	cd ../..
-
-local-run: lint build-local
-	./dist/service
-
-debug-service: deps lint
-	go run cmd/service-user/main.go
-
 lint:
 	golangci-lint run
 
@@ -23,38 +13,42 @@ test: deps
 deps:
 	wire ./...
 
+sanitize: deps lint test
+
 local-setup:
 	docker run --name=mysql-server -d -p 3306:3306 mysql/mysql-server:latest
 	echo "mysql server initialized"
-	docker run -d --name redis-stack -p 6379:6379 -p 8001:8001 redis/redis-stack:latest
+	docker run --name=redis-stack -d -p 6379:6379 -p 8001:8001 redis/redis-stack:latest
 	echo "redis initialized"
+	docker run --name=aws-localstack -d -p 4566:4566 -p 4571:4571 localstack/localstack
+	echo "localstack initialized"
 
-sanitize: deps lint test
 
 local-docker-up:
-	docker start mysql-server redis-stack
-	echo "waiting for mysql and redis to be ready..."
-	sleep 1s
+	docker start mysql-server redis-stack aws-localstack
+	echo "waiting for mysql, redis, S3 & dynamoDB to be ready..."
+	sleep 10s
 	docker start user-service
 
 local-docker-down:
-	docker stop mysql-server redis-stack user-service
+	docker stop mysql-server redis-stack aws-localstack
 
 local-docker-delete-service:
 	docker stop user-service
 	docker rm user-service
 
-local-redeploy: local-docker-delete-service run-service local-docker-up
+local-redeploy: local-docker-delete-service run-service
 
-local-clean-run: run-service local-docker-up
+local-deploy: local-setup run-service
 
 service-logs:
 	docker logs user-service -f --tail 100
 
 localstack-up:
-	docker run --rm --name=localstack -p 4566:4566 -p 4571:4571 localstack/localstack
+	docker run --name=aws-localstack -d -p 4566:4566 -p 4571:4571 localstack/localstack
+	echo "localstack initialized"
 
-local-dynamo-setup: localstack-up local-aws-setup
+local-dynamo-setup: local-aws-setup
 	aws --endpoint-url=http://localhost:4566 dynamodb create-table \
         --table-name user-images \
         --attribute-definitions \
@@ -81,3 +75,12 @@ local-aws-setup:
 	aws configure set region eu-central-1
 	sleep 1s
 
+debug-build-local:
+	GOOS=linux GOARCH=arm64/v8 cd ./cmd/service-user && go build -a -o ../../dist/service
+	cd ../..
+
+debug-run: lint build-local
+	./dist/service
+
+debug-service: deps lint
+	go run cmd/service-user/main.go
