+++ interfaceadapters/controllers/usercontroller_test.go
@@ -1,4 +1,4 @@
-package usercontroller
+package controllers
 
 import (
 	"encoding/json"
